Cleanup I18N-Module - merge LanguageIdentificationService in LanguageService - I18N more robust, include fallback on error (default or key)
