+++ Engine/Modules/I18n/.phpstorm/.phpstorm.meta.php
@@ -4,9 +4,8 @@ namespace PHPSTORM_META {
 
     if (function_exists('override')) {
         override(\Oforge\Engine\Modules\Core\Manager\Services\ServiceManager::get(0), map([
-            'i18n'                => \Oforge\Engine\Modules\I18n\Services\InternationalizationService::class,
-            'languages'           => \Oforge\Engine\Modules\I18n\Services\LanguageService::class,
-            'language.identifier' => \Oforge\Engine\Modules\I18n\Services\LanguageIdentificationService::class,
+            'i18n'          => \Oforge\Engine\Modules\I18n\Services\InternationalizationService::class,
+            'i18n.language' => \Oforge\Engine\Modules\I18n\Services\LanguageService::class,
         ]));
     }
