+++ Engine/Modules/I18n/Helper/I18N.php
@@ -2,9 +2,10 @@
 
 namespace Oforge\Engine\Modules\I18n\Helper;
 
+use Exception;
 use Oforge\Engine\Modules\Core\Exceptions\ServiceNotFoundException;
 use Oforge\Engine\Modules\I18n\Services\InternationalizationService;
-use Oforge\Engine\Modules\I18n\Services\LanguageIdentificationService;
+use Oforge\Engine\Modules\I18n\Services\LanguageService;
 
 /**
  * Class I18N
@@ -25,19 +26,24 @@ class I18N {
      * @param string|null $language
      *
      * @return string
-     * @throws ServiceNotFoundException
      */
     public static function translate(string $key, ?string $defaultValue = null, ?string $language = null) : string {
-        if (!isset($language)) {
-            self::initCurrentLanguage([]);
-            $language = self::$language;
-        }
-        if (!isset(self::$i18nService)) {
-            /** @var InternationalizationService $service */
-            self::$i18nService = Oforge()->Services()->get('i18n');
+        try {
+            if (!isset($language)) {
+                self::initCurrentLanguage([]);
+                $language = self::$language;
+            }
+            if (!isset(self::$i18nService)) {
+                /** @var InternationalizationService $service */
+                self::$i18nService = Oforge()->Services()->get('i18n');
+            }
+
+            return self::$i18nService->get($key, $language, $defaultValue);
+        } catch (Exception $exception) {
+            Oforge()->Logger()->get()->error($exception->getMessage(), $exception->getTrace());
         }
 
-        return self::$i18nService->get($key, $language, $defaultValue);
+        return self::translateFallback($key, $defaultValue);
     }
 
     /**
@@ -48,12 +54,17 @@ class I18N {
      * @param string|null $defaultValue
      *
      * @return string
-     * @throws ServiceNotFoundException
      */
     public static function twigTranslate($context, string $key, ?string $defaultValue = null) : string {
-        self::initCurrentLanguage($context);
+        try {
+            self::initCurrentLanguage($context);
 
-        return self::translate($key, $defaultValue, self::$language);
+            return self::translate($key, $defaultValue, self::$language);
+        } catch (Exception $exception) {
+            Oforge()->Logger()->get()->error($exception->getMessage(), $exception->getTrace());
+        }
+
+        return self::translateFallback($key, $defaultValue);
     }
 
     /**
@@ -65,11 +76,23 @@ class I18N {
      */
     protected static function initCurrentLanguage($context) {
         if (!isset(self::$language)) {
-            /**@var LanguageIdentificationService $languageIdentificationService */
-            $languageIdentificationService = Oforge()->Services()->get('language.identifier');
+            /**@var LanguageService $languageService */
+            $languageService = Oforge()->Services()->get('i18n.language');
 
-            self::$language = $languageIdentificationService->getCurrentLanguage($context);
+            self::$language = $languageService->getCurrentLanguage($context);
         }
+    }
+
+    /**
+     * Fallback, if translate fails because of errors.
+     *
+     * @param string $key
+     * @param string|null $defaultValue
+     *
+     * @return string
+     */
+    protected static function translateFallback(string $key, ?string $defaultValue = null) : string {
+        return isset($defaultValue) ? $defaultValue : $key;
     }
 
 }
